test: merged config test files
